refactor code and auto-generate sync methods
